+++ core_client/base/api/v3_cluster_node_put_state.py
@@ -33,7 +33,7 @@ def _build_request(
 
 def _build_response(response: httpx.Response):
     if response.status_code == 200:
-        response_200 = TypeAdapter(ClusterNodeState).validate_python(response.json())
+        response_200 = response.json()
         return response_200
     else:
         response_error = TypeAdapter(Error).validate_python(response.json())