+++ core_client/base/api/v3_iam_delete_user.py
@@ -4,7 +4,6 @@ from pydantic import TypeAdapter, validate_call
 
 from ...models import Client
 from ..models import Error
-from ..models.v3 import IamUser
 
 
 @validate_call()
@@ -35,7 +34,7 @@ def _build_request(
 
 def _build_response(response: httpx.Response):
     if response.status_code == 200:
-        response_200 = TypeAdapter(IamUser).validate_python(response.json())
+        response_200 = response.json()
         return response_200
     else:
         response_error = TypeAdapter(Error).validate_python(response.json())